requirements.txt
# requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
Pillow==9.5.0
